requirements.txt
affine==2.4.0
annotated-types==0.7.0
attrs==24.2.0
autopep8==2.3.1
boto3==1.35.4
botocore==1.35.4
cads-api-client==1.2.0
cdsapi==0.7.1
certifi==2024.7.4
cfgv==3.4.0
cftime==1.6.4
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
configobj==5.0.8
contourpy==1.2.1
cycler==0.12.1
Deprecated==1.2.14
distlib==0.3.8
donfig==0.8.1.post1
eotile==0.2.8
filelock==3.15.4
fiona==1.9.6
fonttools==4.53.1
GDAL @ file:///home/conda/feedstock_root/build_artifacts/libgdal-core_1723837536297/work/build/swig/python
geographiclib==2.0
geopandas==0.9.0
geopy==2.4.1
h5netcdf==1.3.0
h5py==3.11.0
identify==2.6.0
idna==3.7
imageio==2.35.1
importlib_resources==6.4.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.43.0
loguru==0.7.2
lxml==5.2.2
matplotlib==3.9.2
morecantile==5.4.0
multiurl==0.3.1
netCDF4==1.7.1.post2
networkx==3.3
nodeenv==1.9.1
numba==0.60.0
numexpr==2.10.1
numpy==2.0.1
openeo==0.31.0
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.8.0
psutil==6.0.0
pycodestyle==2.12.1
pydantic==2.8.2
pydantic_core==2.20.1
pykdtree==1.3.12
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
pyresample==1.29.0
pystac==1.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
rasterio==1.3.10
requests==2.32.3
requests-toolbelt==1.0.0
richdem @ file:///home/conda/feedstock_root/build_artifacts/richdem_1716430392807/work/wrappers/pyrichdem
rio-cogeo==5.3.3
rioxarray==0.17.0
ruff==0.5.0
s3transfer==0.10.2
satio==1.1.17
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
Shapely==1.8.5.post1
six==1.16.0
snuggs==1.4.7
statsmodels==0.14.2
threadpoolctl==3.5.0
tifffile==2024.8.10
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
utm==0.7.0
virtualenv==20.26.3
wasdi==0.8.6.3
wrapt==1.16.0
xarray==2024.3.0
